requirements.txt
ivis==2.0.9
jax==0.4.14
jaxlib==0.4.14
jaxopt==0.8
lolP==0.0.4
matplotlib==3.7.1
numpy==1.24.3
pandas==2.0.2
plotly==5.15.0
pyarrow==12.0.0
Requests==2.31.0
scikit-learn==1.2.2
scipy==1.12.0
seaborn==0.12.2
setuptools==67.7.2
tensorflow==2.15.0
xgboost==2.0.3
tqdm==4.65.0
Jinja2==3.1.4
